# openapi.yml
openapi: 3.0.0
info:
  title: EduMap API
  description: API для интерактивной карты учебных курсов
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Описание ошибки
      required:
        - detail
...
    UserRole:
      type: string
      enum: [student, teacher, admin]
    AccountStatus:
      type: string
      enum: [created, confirmed, blocked]
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "student@example.com"
        password:
          type: string
          format: password
          minLength: 6
          maxLength: 128
          example: "securepass123"
        first_name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Иван"
        last_name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Петров"
      required:
        - email
        - password
        - first_name
        - last_name
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "student@example.com"
        password:
          type: string
          format: password
          minLength: 6
          maxLength: 128
          example: "securepass123"
      required:
        - email
        - password
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
      required:
        - access_token
        - refresh_token
        - token_type
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: "bearer"
      required:
        - access_token
        - refresh_token
        - token_type
    LogoutResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    MeResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/AccountStatus'
      required:
        - id
        - email
        - first_name
        - last_name
        - role
        - status
    VerifyAccountRequest:
      type: object
      properties:
        code:
          type: string
          format: uuid
      required:
        - code
    ChangePasswordRequest:
      type: object
      properties:
        old_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
          minLength: 6
          maxLength: 128
        confirm_password:
          type: string
          format: password
          minLength: 6
          maxLength: 128
      required:
        - new_password
        - confirm_password
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    ResetPasswordRequest:
      type: object
      properties:
        code:
          type: string
          format: uuid
        new_password:
          type: string
          format: password
          minLength: 6
          maxLength: 128
        confirm_password:
          type: string
          format: password
          minLength: 6
          maxLength: 128
      required:
        - code
        - new_password
        - confirm_password
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required:
        - message
        - success
...
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email уже зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/verify:
    post:
      tags: [auth]
      summary: Подтверждение аккаунта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyAccountRequest'
      responses:
        '200':
          description: Аккаунт подтвержден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Код недействителен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Код не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [auth]
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "refresh_token=...; HttpOnly; Secure; SameSite=Lax"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Неверный email или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Обновление токена
      parameters:
        - in: cookie
          name: refresh_token
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Токен обновлен
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Невалидный refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags: [auth]
      summary: Выход из системы
      parameters:
        - in: cookie
          name: refresh_token
          schema:
            type: string
      responses:
        '200':
          description: Успешный выход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags: [auth]
      summary: Получить информацию о текущем пользователе
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Запрос на восстановление пароля
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Инструкция отправлена на email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Сброс пароля
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Пароль успешно изменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Код не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/change-password:
    post:
      tags: [auth]
      summary: Изменение пароля
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Пароль успешно изменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...